Prevent access to todo table and functions unless user logged in

+++ js/controller.js
@@ -86,39 +86,43 @@ angular.module('RouteControllers', [])
         $scope.authToken = store.get('authToken');
         $scope.username = store.get('username');
 
-        $scope.todos = {};
+        if (!store.get('authToken')) {
+            $location.path("/accounts/register");
+        } else {
+            $scope.todos = {};
 
-        TodoAPIService.getTodos(url + "todo/", $scope.username, $scope.authToken).then(function(results) {
-            $scope.todos = results.data;
-            console.log($scope.todos);
-        }).catch(function(err) {
-            console.log(err);
-        });
-
-        $scope.submitForm = function() {
-            if ($scope.todoForm.$valid) {
-                $scope.todo.username = $scope.username;
-                $scope.todos.push($scope.todo);
-
-                TodoAPIService.createTodo(url + "todo/", $scope.todo, $scope.authToken).then(function(results) {
-                    console.log(results);
-                }).catch(function(err) {
-                    console.log(err);
-                });
-            }
-        };
-
-        $scope.editTodo = function(id) {
-            $location.path("/todo/edit/" + id);
-        };
-
-        $scope.deleteTodo = function(id) {
-            TodoAPIService.deleteTodo(url + "todo/" + id, $scope.username, $scope.authToken).then(function(results) {
-                console.log(results);
-            }).catch(function(err) {
+            TodoAPIService.getTodos(url + "todo/", $scope.username, $scope.authToken).then(function (results) {
+                $scope.todos = results.data;
+                console.log($scope.todos);
+            }).catch(function (err) {
                 console.log(err);
             });
-        };
+
+            $scope.submitForm = function () {
+                if ($scope.todoForm.$valid) {
+                    $scope.todo.username = $scope.username;
+                    $scope.todos.push($scope.todo);
+
+                    TodoAPIService.createTodo(url + "todo/", $scope.todo, $scope.authToken).then(function (results) {
+                        console.log(results);
+                    }).catch(function (err) {
+                        console.log(err);
+                    });
+                }
+            };
+
+            $scope.editTodo = function (id) {
+                $location.path("/todo/edit/" + id);
+            };
+
+            $scope.deleteTodo = function (id) {
+                TodoAPIService.deleteTodo(url + "todo/" + id, $scope.username, $scope.authToken).then(function (results) {
+                    console.log(results);
+                }).catch(function (err) {
+                    console.log(err);
+                });
+            };
+        }
     })
 
     .controller('EditTodoController', function($scope, $location, $routeParams, TodoAPIService, store) {
